requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
compress-pickle==2.1.0
Flask==2.2.3
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kafka-python==2.0.2
MarkupSafe==2.1.2
numpy==1.24.2
pandas==1.5.3
prometheus-client==0.16.0
python-dateutil==2.8.2
pytz==2023.2
requests==2.28.2
scikit-learn==1.2.2
scikit-surprise==1.1.3
scipy==1.10.1
six==1.16.0
surprise==0.1
threadpoolctl==3.1.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
